\name{TianHaoR}
\alias{TheStart}
\title{Hello, R!}
\usage{
hao.test()
tang.test()
lv.test()
yang.test()
xia.test()
pcascore()
fcascore()
}
\description{
The mean value and covariance test methods for multiple normal vectors are provided, and the calculation methods of principal component score and factor score in principal component analysis are provided
}
\examples{
hao.test(a,miu,cov,alpha)
tang.test(a,miu)
lv.test(x1,x2,alpha)
yang.test(x1,x2,alpha)
xia.test(a,sigma0,alpha)
pcascore(a)
fcascore(a)
}
